Camera3: Protect against simultaneous camera instances
When frameworks try to open the 2nd camera instance, CTS test
allows open to fail, but doesn't allow startPreview to fail.

The way camera daemon is written, ISP hardware is reserved during
startPreview, but only released when all streams are removed.
The removal of streams happens when camera is closed. Because of this,
the current logic allows the 2nd open to succeed, but startPreview
will fail. This is not good!

Block the 2nd open in HAL until we have a better underlying design
to address this issue.

Change-Id: If00b1227f0794171cea3b43c7af28dd766a1943e
Signed-off-by: Iliyan Malchev <malchev@google.com>