Camera3: Only queue snapshot YUV buffers before snapshot request
Since we are starting all configured streams before the very first
request, we want to make sure we are only queuing the YUV buffers
when we receive the snapshot request.

This logic only applies to non-reprocessing capture. For reprocessing
capture, since we already have input buffer, we don't need to queue
any buffer into hardware.

Bug: 10544282

Change-Id: Ia3434160f4944649fd327b839f9130edb32b54d1