QCamera3: Support HDR+ processing in HDR+ service
When HDR+ mode is enabled (persist.camera.hdrplus = 1), request
an additional RAW buffers for every 10 capture requests. Send
the RAW buffer to HDR+ service as input buffers. Also send
result metadata to HDR+ service for every frame.

In processCaptureRequest, decide if a capture request is a HDR+
request by checking the capture intent and post processing quality.
For a HDR+ capture request, get an YUV buffer from pic channel
and send the YUV buffer to HDR+ service to be filled with HDR+
processed YUV data. When the YUV buffer is filled, return the
YUV buffer to pic channel for JPEG encoding.

Refactored the logic to send shutter callbacks and results so that
regular, HDR+, and reprocess shutter callbacks and results are
in the order of increasing frame numbers.

Bug: 28637032
Change-Id: Id5829dc8440fa56574f9b5f0fde7b2d0d694e14e