QCamera3: Fix jpeg settings for HDR+ requests
Save the capture request settings for pending HDR+ requests and
use the settings to override HDR+ capture results given by HDR+
client. This is necessary for the encoder to get the right jpeg
settings because HDR+ capture results given by HDR+ client belong
to ZSL frames, requested in the past.

Test: camera2hdr with HDR+ enabled
Bug: 28637032
Change-Id: I44fed28e244e41823a2c8a73587e0172a003eb85