QCamera2: HAL3: Use focus distance only for AF off
Focus distance needs to be included in the default request template to
pass CTS.
However, this means the app might send a default focus distance value
even if it does not intend to use manual focus.
HAL must filter this out for all non-manual AF modes to avoid complaints
from the backend.

CRs-fixed: 2012397
Change-Id: I78bc3ca9806ee02bd87eef0dbc52e04a6c38b22c