QCamera2: Allocate depth buffers dynamically
'depth_data' is always allocated as part of the metadata
buffers. This is wasting memory in all cases where the
depth channel is not part of the configured streams.
To avoid this, allocate depth data buffers dynamically
ony when a depth stream is present in the configuration.

Bug: 38292566
Test: Manual using application,
Complete Camera/Camera2 CTS.
Change-Id: I1fbdb263551fa6fb0222346e0f5bbc3415ae8f2f