QCamera2: HAL3: Preview TNR logic
Preview TNR is disabled by default but should be enabled under specific
circumstances.
Previous logic was placed after channel creation which rendered it moot.
This change also adds the condition that even if EIS3 is used, preview
TNR should still be enabled for front camera if the other conditions are
satisfied (front camera does not use EIS3 regardless of setprop).

Test: Camera CTS
Bug: 62863294
Change-Id: Ia6cc78273ae79830d974a997ac00011f0b022407