QCamera2: HAL3: Fix null AF_MODE metadata issue in HFR
In HFR case, if the first metadata doesn't map to any partial frame
number, current logic skip all metadata entries and sends empty result
up to app.

This conflicts with HAL jumpstart in the sense that HAL uses first
metadata as the partial result for the very first request. As a result,
AF_MODE metadata becomes null.

The fix is not to skip the partial result metadata for first request.

Test: Camera CTS
Bug: 64207183
Change-Id: I4643882ff504cee3ce3a344417c3ee48d9488677