QCamera2: HAL3: Add scene distance metadata
Scene distance is a new vendor tag added to indicate HAL's estimation of
the scene distance. This is made possible if the camera device is
equipped with laser.

The value is in the units of mm. If the camera device cannot estimate
the distance, or it doesn't have laser to make such estimation, the
value is -1.

Bug: 65179074
Test: Camera CTS
Change-Id: I378cbcd0a65e58452bb59e956694345fb9202198