QCamera3: Make sure stream IDs are unique
If the stream configuration is called again, the indices of
the stream in new stream configuration may be the same as
the old stream configuration, which results in the same
stream ID for two different streams after stream cleanup.
This CL ensures that stream IDs are unique after each
stream configuration.

Test: CTS
Bug: 64117056
Change-Id: I1dd7d8e67a49b48d6f6448d1ed37b1537555b6e7