Probe powerHAL HIDL first
Right now AIDL doesn't provide an API for vendor code to query if the
AIDL HAL declared. This will lead to 1) long blocking call for device
doesn't have the AIDL HAL and also 2) potential problem if the service
starts late after 5s timeout. This is a WAR to change the order to
prevent long blocking call on devices having Hidl service.

Bug: 150183391
Test: Build
Change-Id: I1966a094d422540f314537ffd26b4498d656af6a